requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
pydantic==2.11.3
python-dotenv==1.1.0
celery==5.5.1
redis==5.2.1
pdfplumber==0.11.6
python-docx==1.1.2
pillow==11.2.1
pytesseract==0.3.13
spacy==3.8.7
transformers==4.51.3
torch==2.7.0
